+++ packages/core/src/components/skiplink/skiplink.interfaces.ts
@@ -1,0 +1,5 @@
+export interface SkiplinkClickEvent {
+  originalEvent: MouseEvent;
+  /** True when user clicked the card while holding Ctrl, Alt or other modifiers, or when the card is right-clicked. Can be used to determine navigation. */
+  isModifiedEvent: boolean;
+}