+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## Next
 
+### Added
+* Skiplink: New component ([#2746](https://github.com/dso-toolkit/dso-toolkit/issues/2746))
+
 ## ↩️ Release 65.1.1 - 2024-11-14
 
 ### Removed